+++ packages/did-ui-react/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.1-alpha.3](https://github.com/Portkey-Wallet/portkey-web/compare/v1.5.0...v1.5.1-alpha.3) (2023-11-01)
+
+### Features
+
+- 🎸 checkWalletSecurity ([d16bd86](https://github.com/Portkey-Wallet/portkey-web/commit/d16bd86769c6e2df76e966f147af39fe8fec08ec))
+
+### Bug Fixes
+
+- 🐛 add loading - managerApprove ([303deaa](https://github.com/Portkey-Wallet/portkey-web/commit/303deaa169e599acb766677e2603d84108560d60))
+- 🐛 guardian origin chainId ([057cc99](https://github.com/Portkey-Wallet/portkey-web/commit/057cc99c7bf5ff7713c873d1ce3fd33a93a94af1))
+
+## [1.5.1-alpha.2](https://github.com/Portkey-Wallet/portkey-web/compare/v1.5.1-alpha.1...v1.5.1-alpha.2) (2023-11-01)
+
+### Bug Fixes
+
+- 🐛 add loading - managerApprove ([9aff0a9](https://github.com/Portkey-Wallet/portkey-web/commit/9aff0a9a182022875aafdcfe1d1144b4493fd6c0))
+
+## [1.5.1-alpha.1](https://github.com/Portkey-Wallet/portkey-web/compare/v1.5.1-alpha.0...v1.5.1-alpha.1) (2023-11-01)
+
+**Note:** Version bump only for package @portkey/did-ui-react
+
+## [1.5.1-alpha.0](https://github.com/Portkey-Wallet/portkey-web/compare/v1.5.0...v1.5.1-alpha.0) (2023-11-01)
+
+### Features
+
+- 🎸 checkWalletSecurity ([b2ced57](https://github.com/Portkey-Wallet/portkey-web/commit/b2ced5726920fec942012b30be15152dc4117f9a))
+
 ## [1.5.0](https://github.com/Portkey-Wallet/portkey-web/compare/v1.5.0-alpha.23...v1.5.0) (2023-10-30)
 
 ### Bug Fixes